Move get_aggregated_signature from tauri client to coconut interface
